[nemo-qml-plugin-contacts] Use libcontacts
SeasideCache has been extracted to libcontacts.
