+++ tests/basename.pri
@@ -1,0 +1,2 @@
+equals(QT_MAJOR_VERSION, 4): BASENAME=nemo-qml-plugins
+equals(QT_MAJOR_VERSION, 5): BASENAME=nemo-qml-plugins-qt5
