+++ tests/tests.xml.in
@@ -1,14 +1,14 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <testdefinition version="1.0">
-   <suite name="contacts-tests" domain="mw">
+   <suite name="@BASENAME@-contacts-tests" domain="mw">
        <description>Contacts QML plugin automatic tests</description>
-       <set name="unit-tests" feature="QML Contacts">
+       <set name="@BASENAME@-contacts-test0" feature="QML Contacts">
            <description>Contacts QML plugin automatic tests</description>
            <case manual="false" name="seasideperson">
-               <step>/opt/tests/nemo-qml-plugins-qt5/contacts/tst_seasideperson</step>
+               <step>/opt/tests/@BASENAME@/contacts/tst_seasideperson</step>
            </case>
            <case manual="false" name="seasidefilteredmodel">
-               <step>/opt/tests/nemo-qml-plugins/contacts/tst_seasidefilteredmodel</step>
+               <step>/opt/tests/@BASENAME@/contacts/tst_seasidefilteredmodel</step>
            </case>
        </set>
    </suite>
