+++ tests/tst_seasidefilteredmodel/synchronizelists.h
@@ -1,0 +1,7 @@
+// Include the real header (though we haven't configured the include path in this project)
+#include <QtGlobal>
+#if QT_VERSION >= QT_VERSION_CHECK(5, 0, 0)
+#include <contactcache-qt5/synchronizelists.h>
+#else
+#include <contactcache/synchronizelists.h>
+#endif
