+++ tests/tst_seasideperson/tst_seasideperson.pro
@@ -1,4 +1,11 @@
 include(../common.pri)
+
 TARGET = tst_seasideperson
 
-SOURCES += tst_seasideperson.cpp
+equals(QT_MAJOR_VERSION, 4): PKGCONFIG += contactcache
+equals(QT_MAJOR_VERSION, 5): PKGCONFIG += contactcache-qt5
+
+SOURCES += $$SRCDIR/seasideperson.cpp \
+           tst_seasideperson.cpp
+
+HEADERS += $$SRCDIR/seasideperson.h
