--- tests/tst_synchronizelists/tst_synchronizelists.pro
@@ -1,4 +1,0 @@
-include(../common.pri)
-TARGET = tst_synchronizelists
-
-SOURCES += tst_synchronizelists.cpp
